scraping1/actividad1/webscrapping-server-main/requirements.txt
APScheduler==3.10.4
fastapi==0.115.8
psycopg2==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
python-telegram-bot==21.10
Requests==2.32.3
starlette==0.45.3
uvicorn==0.34.0
